% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_plan}
\alias{convert_plan}
\title{Convert files from a conversion plan}
\usage{
convert_plan(
  plan,
  bids_dir,
  clobber = FALSE,
  bids_uri = "URI",
  post_op = list()
)
}
\arguments{
\item{plan}{Plan from \code{\link[=plan_conversion]{plan_conversion()}}}

\item{bids_dir}{Destination BIDS directory}

\item{clobber}{Overwrite existing files}

\item{bids_uri}{IntendedFor path style: URI or relative}

\item{post_op}{Optional list of post-op command objects}
}
\value{
Conversion summary list
}
\description{
Convert files from a conversion plan
}
