% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm2niix.R
\name{dcm2niix_version}
\alias{dcm2niix_version}
\title{Query dcm2niix version}
\usage{
dcm2niix_version(binary = "dcm2niix")
}
\arguments{
\item{binary}{Name or path of dcm2niix binary}
}
\value{
Version text
}
\description{
Query dcm2niix version
}
