% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{dcmtobids_main}
\alias{dcmtobids_main}
\title{CLI entrypoint}
\usage{
dcmtobids_main(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{args}{Character vector of CLI args}
}
\value{
Process status code
}
\description{
CLI entrypoint
}
