% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm2niix.R
\name{find_dcm2niix}
\alias{find_dcm2niix}
\title{Find dcm2niix binary}
\usage{
find_dcm2niix(binary = "dcm2niix")
}
\arguments{
\item{binary}{Name or path of dcm2niix binary}
}
\value{
Absolute path to executable
}
\description{
Find dcm2niix binary
}
