% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cli.R
\name{install_cli}
\alias{install_cli}
\title{Install dcmtobids CLI launcher in PATH}
\usage{
install_cli(install_dir = fs::path_abs("~/.local/bin"), overwrite = TRUE)
}
\arguments{
\item{install_dir}{Directory to place launcher script}

\item{overwrite}{Overwrite existing launcher}
}
\value{
Path to installed launcher
}
\description{
Install dcmtobids CLI launcher in PATH
}
