% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{plan_conversion}
\alias{plan_conversion}
\title{Build a conversion plan}
\usage{
plan_conversion(
  config,
  input_dir,
  participant_label,
  session_label = "",
  search_method = config$search_method \%||\% "fnmatch",
  case_sensitive = config$case_sensitive \%||\% TRUE,
  auto_extract_entities = config$auto_extract_entities \%||\% FALSE,
  do_not_reorder_entities = config$do_not_reorder_entities \%||\% FALSE,
  dup_method = config$dup_method \%||\% "run",
  warning_sink = NULL
)
}
\arguments{
\item{config}{Parsed config list}

\item{input_dir}{Directory containing dcm2niix sidecars}

\item{participant_label}{Participant ID (with or without sub- prefix)}

\item{session_label}{Session ID (with or without ses- prefix)}

\item{search_method}{Matching method: fnmatch or re}

\item{case_sensitive}{Whether matching should be case sensitive}

\item{do_not_reorder_entities}{Keep entity order as given}

\item{dup_method}{Duplicate strategy: run or dup}
}
\value{
Data frame plan
}
\description{
Build a conversion plan
}
