% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{print_dry_run_plan}
\alias{print_dry_run_plan}
\title{Print dry-run plan}
\usage{
print_dry_run_plan(plan, bids_dir = NULL)
}
\arguments{
\item{plan}{Plan from \code{\link[=plan_conversion]{plan_conversion()}}}

\item{bids_dir}{Optional destination root for rendering destination paths}
}
\value{
Invisibly returns table data frame
}
\description{
Print dry-run plan
}
