% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Read a dcmtobids config JSON}
\usage{
read_config(config_path)
}
\arguments{
\item{config_path}{Path to JSON config}
}
\value{
Parsed config list
}
\description{
Read a dcmtobids config JSON
}
