% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{summarize_population}
\alias{summarize_population}
\title{Build structured summary report from an IntendedFor population run}
\usage{
summarize_population(population)
}
\arguments{
\item{population}{Summary from \code{\link[=populate_intended_for_plan]{populate_intended_for_plan()}}}
}
\value{
Named list report
}
\description{
Build structured summary report from an IntendedFor population run
}
