% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm2niix.R
\name{run_dicom_conversion}
\alias{run_dicom_conversion}
\title{Run full DICOM to BIDS pipeline}
\usage{
run_dicom_conversion(
  config_path,
  dicom_dirs,
  bids_dir,
  participant_label,
  session_label = "",
  auto_extract_entities = FALSE,
  do_not_reorder_entities = FALSE,
  dcm2niix_bin = "dcm2niix",
  dcm2niix_options = DEFAULT_DCM2NIIX_OPTIONS,
  force_dcm2bids = FALSE,
  skip_dcm2niix = FALSE,
  clobber = FALSE
)
}
\arguments{
\item{config_path}{Path to dcmtobids config JSON}

\item{dicom_dirs}{Character vector of DICOM directories or archives}

\item{bids_dir}{Destination BIDS directory}

\item{participant_label}{Participant label}

\item{session_label}{Session label}

\item{dcm2niix_bin}{Name/path of dcm2niix binary}

\item{dcm2niix_options}{dcm2niix option string}

\item{force_dcm2bids}{Recreate helper output before conversion}

\item{skip_dcm2niix}{Skip dcm2niix and copy NIfTI/JSON inputs}

\item{clobber}{Overwrite existing destination files}
}
\value{
List containing helper and conversion results
}
\description{
Run full DICOM to BIDS pipeline
}
